Move relinking from second step to third step
Some explosions can have loopback dependencies: A block has a dependency outside the explosion and then a block inside the explosion is dependent on that. If we link in stage two, this outside dependency has not been added yet so the block dependent on it becomes independent and might be placed before it. Since the relinking stage doesn't affect external dependency checking, this shouldn't cause any regressions.